requirements.txt
# Core dependencies
pyyaml>=6.0
click>=8.0

# Optional: For AI-powered interview (if implemented)
# openai>=1.0.0
# anthropic>=0.3.0

